feat(today): host the Sleep-debt-ledger sleep card (#today-hosted-cards)
Card 3 on the shared SleepModel foundation, mirroring cards 1-2. Lift
SleepView.sleepDebtLedger + debtDeltaBars + debt-only formatters into a
standalone SleepDebtLedgerCard(model:) reading model.sleepDebtLedger (the shared
nap-credited ledger - not recomputed). Register HostedCard.sleepDebt
(byte-identical sleep.sleepDebt id, existing localized title); reuse the card-1
Today-host plumbing. Android SleepDebtLedgerHostCard + wiring; its debt
formatters were already shared in SleepFormatting.kt. Dropped a dead single-arg
debtHeadline wrapper picked up in the lift.

Android compile + i18n green locally; Swift via gate.
